+++ packages/spar/CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Initial release of @turkish-technology/glide
+- Initial release of @turkish-technology/spar
 - Headless React components with TypeScript support
 - Components:
   - Accordion